Replace expr version of PS18 with new version
